Fix reverse iterator causing undefined behaviour
Calling the encoding's prev_code_point function will try to read
code units before the pointed to value, which is undefined behaviour
when called with the pointer to the beginning of a string.

Fixed by storing the pointer to the last code point
in the reverse iterator, and using a special value for the end.